+++ docs/pages/components/calendar/fragments/week-start.md
@@ -8,15 +8,15 @@ const App = () => {
   const [isoWeek, setIsoWeek] = React.useState(false);
   const [showWeekNumbers, setShowWeekNumbers] = React.useState(true);
   return (
-    <HStack divider={<Divider vertical />} spacing={40} style={{ height: 410 }}>
-      <HStack style={{ width: 340 }}>
-        <Calendar
-          compact
-          weekStart={weekStart}
-          showWeekNumbers={showWeekNumbers}
-          isoWeek={isoWeek}
-        />
-      </HStack>
+    <HStack wrap divider={<Divider vertical />} spacing={10} style={{ height: 320 }}>
+      <Calendar
+        compact
+        weekStart={weekStart}
+        showWeekNumbers={showWeekNumbers}
+        isoWeek={isoWeek}
+        style={{ width: 320 }}
+      />
+
       <VStack spacing={20}>
         <RadioGroup
           inline
