+++ src/Calendar/MonthDropdown/index.tsx
@@ -1,0 +1,3 @@
+import MonthDropdown from './MonthDropdown';
+
+export default MonthDropdown;
