+++ src/Calendar/styles/index.less
@@ -96,6 +96,7 @@
   &-panel &-month-dropdown {
     margin-left: 10px;
     margin-right: 10px;
+    border-radius: 6px;
     width: ~'calc(100% - 20px)';
   }
 
@@ -498,15 +499,11 @@
   }
 
   &-list {
-    display: block;
-    min-width: 30px * 6 + 4px * 5;
-    .clearfix();
+    display: grid;
+    grid-template-columns: repeat(4, 25%);
   }
 
   &-cell {
-    display: inline-block;
-    float: left;
-    width: ~'calc((100% - @{calendar-table-cell-content-today-border-width} * 12 ) / 6)';
     margin: 1px;
     text-align: center;
     vertical-align: middle;
@@ -519,8 +516,6 @@
 
   &-cell-content {
     .picker-calendar-cell-content();
-    width: @calendar-in-menu-content-side-length;
-    height: @calendar-in-menu-content-side-length;
   }
 
   &-cell:not(&-cell-active).disabled &-cell-content {
