+++ src/Calendar/utils/index.ts
@@ -1,1 +1,2 @@
 export { getAriaLabel } from './getAriaLabel';
+export { isEveryDateInMonth } from './isEveryDateInMonth';
