+++ src/DatePicker/DatePicker.tsx
@@ -2,10 +2,11 @@ import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import mapValues from 'lodash/mapValues';
 import pick from 'lodash/pick';
-import IconCalendar from '@rsuite/icons/legacy/Calendar';
+import IconCalendar from '@rsuite/icons/Calendar';
 import TimeIcon from '@rsuite/icons/Time';
 import CalendarContainer from '../Calendar/CalendarContainer';
 import { useCalendarDate } from '../Calendar/hooks';
+import { isEveryDateInMonth } from '../Calendar/utils';
 import Toolbar, { RangeType } from './Toolbar';
 import Stack from '../Stack';
 import PredefinedRanges from './PredefinedRanges';
@@ -14,7 +15,6 @@ import InputGroup from '../InputGroup';
 import useMonthView from './hooks/useMonthView';
 import useFocus from './hooks/useFocus';
 import useCustomizedInput from './hooks/useCustomizedInput';
-import { isEveryDateInMonth } from '../Calendar/MonthDropdown';
 import { mergeRefs, partitionHTMLProps, createChainedFunction } from '@/internals/utils';
 import {
   useClassNames,
