+++ src/DateRangePicker/styles/index.less
@@ -56,10 +56,6 @@
 
   .rs-calendar-month-dropdown {
     z-index: @zindex-date-range-picker-calendar-dropdown;
-
-    &-list {
-      width: 185px;
-    }
   }
 
   .rs-picker-daterange-panel-show-one-calendar {
